common: move fork check to EFA provider
The autoconf was done incorrectly causing builds to fail. Move the fork
check function into the EFA provider for now.

Fixes #6327
